+++ packages/framework/test/response_header_test.dart
@@ -41,7 +41,7 @@ void main() {
   }, skip: true);
 
   test('Add Response Header', () async {
-    http.responseHeader({
+    http.addResponseHeader({
       'X-XSRF_TOKEN':
           'a591a6d40bf420404a011733cfb7b190d62c65bf0bcda32b57b277d9ad9f146e'
     });